backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
paho-mqtt==1.6.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0